feat(formatter-html): extract ApiMetadataCard from SummaryCard
Move the stats bar (operations / schemas / tags / security counts) into a
new ApiMetadataCard component. SummaryCard gains a showApiMetadata prop
(default true) that renders ApiMetadataCard conditionally, preserving
existing behavior for all callers.

Refs #342
